+++ ODK.Services/Members/MemberAdminService.cs
@@ -1,12 +1,10 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Linq;
 using System.Threading.Tasks;
 using ODK.Core.Chapters;
 using ODK.Core.Members;
 using ODK.Services.Caching;
 using ODK.Services.Exceptions;
-using ODK.Services.Files;
 
 namespace ODK.Services.Members
 {
@@ -27,7 +25,11 @@ namespace ODK.Services.Members
 
         public async Task DeleteMember(Guid currentMemberId, Guid memberId)
         {
-            Member member = await GetMember(currentMemberId, memberId);
+            Member? member = await GetMember(currentMemberId, memberId);
+            if (member == null)
+            {
+                return;
+            }
 
             await _memberRepository.DeleteMember(member.Id);
 
@@ -37,36 +39,31 @@ namespace ODK.Services.Members
 
         public async Task DisableMember(Guid currentMemberId, Guid id)
         {
-            Member member = await GetMember(currentMemberId, id);
+            Member? member = await GetMember(currentMemberId, id);
+            if (member == null)
+            {
+                return;
+            }
 
             await _memberRepository.DisableMember(member.Id);
         }
 
         public async Task EnableMember(Guid currentMemberId, Guid id)
         {
-            Member member = await GetMember(currentMemberId, id);
+            Member? member = await GetMember(currentMemberId, id);
+            if (member == null)
+            {
+                return;
+            }
 
             await _memberRepository.EnableMember(member.Id);
         }
 
-        public async Task<Member> GetMember(Guid currentMemberId, Guid memberId)
+        public async Task<Member?> GetMember(Guid currentMemberId, Guid memberId)
         {
             return await GetMember(currentMemberId, memberId, false);
         }
-
-        public async Task<CsvFile> GetMemberImportFile(Guid currentMemberId, Guid chapterId)
-        {
-            await AssertMemberIsChapterAdmin(currentMemberId, chapterId);
-
-            IReadOnlyCollection<ChapterProperty> properties = await ChapterRepository.GetChapterProperties(chapterId);
-
-            MemberCsvFileHeaders headers = new MemberCsvFileHeaders(properties);
-
-            CsvFile file = new CsvFile();
-            file.Header.AddValues(headers.ToFields());
-            return file;
-        }
-
+        
         public async Task<IReadOnlyCollection<Member>> GetMembers(Guid currentMemberId, Guid chapterId, bool requireSuperAdmin = false)
         {
             if (requireSuperAdmin)
@@ -81,9 +78,13 @@ namespace ODK.Services.Members
             return await _memberRepository.GetMembers(chapterId, true);
         }
 
-        public async Task<MemberSubscription> GetMemberSubscription(Guid currentMemberId, Guid memberId)
+        public async Task<MemberSubscription?> GetMemberSubscription(Guid currentMemberId, Guid memberId)
         {
-            Member member = await GetMember(currentMemberId, memberId);
+            Member? member = await GetMember(currentMemberId, memberId);
+            if (member == null)
+            {
+                return null;
+            }
 
             return await _memberRepository.GetMemberSubscription(member.Id);
         }
@@ -94,90 +95,24 @@ namespace ODK.Services.Members
 
             return await _memberRepository.GetMemberSubscriptions(chapterId);
         }
-
-        public async Task ImportMembers(Guid currentMemberId, Guid chapterId, CsvFile file)
+        
+        public async Task RotateMemberImage(Guid currentMemberId, Guid memberId, int degrees)
         {
-            await AssertMemberIsChapterAdmin(currentMemberId, chapterId);
-
-            IReadOnlyCollection<ChapterProperty> chapterProperties = await ChapterRepository.GetChapterProperties(chapterId);
-
-            MemberCsvFileHeaders headers = new MemberCsvFileHeaders(chapterProperties);
-
-            IDictionary<string, int> headerIndexes = file.GetColumnIndexes();
-
-            IDictionary<int, Member> members = new Dictionary<int, Member>();
-
-            IDictionary<int, MemberSubscription> membersSubscriptions = new Dictionary<int, MemberSubscription>();
-
-            IDictionary<int, IReadOnlyCollection<MemberProperty>> membersProperties =
-                new Dictionary<int, IReadOnlyCollection<MemberProperty>>();
-
-            for (int i = 0; i < file.Rows.Count; i++)
+            Member? member = await GetMember(currentMemberId, memberId);
+            if (member == null)
             {
-                CsvRow row = file.Rows.ElementAt(i);
-
-                members.Add(i, new Member(Guid.Empty, chapterId,
-                    row.Value(headerIndexes[MemberCsvFileHeaders.Email]),
-                    true,
-                    row.Value(headerIndexes[MemberCsvFileHeaders.FirstName]),
-                    row.Value(headerIndexes[MemberCsvFileHeaders.LastName]),
-                    DateTime.UtcNow,
-                    true,
-                    false,
-                    0));
-
-                if (DateTime.TryParse(row.Value(headerIndexes[MemberCsvFileHeaders.SubscriptionExpiry]), out DateTime subscriptionExpiry))
-                {
-                    membersSubscriptions.Add(i, new MemberSubscription(Guid.Empty, SubscriptionType.Full, subscriptionExpiry));
-                }
-
-                membersProperties.Add(i, chapterProperties
-                    .Select(x => new MemberProperty(Guid.Empty, Guid.Empty, x.Id, row.Value(headerIndexes[x.Label])))
-                    .ToArray());
+                return;
             }
 
-            // TODO: validate members
-
-            for (int i = 0; i < file.Rows.Count; i++)
-            {
-                Guid memberId = await _memberRepository.CreateMember(members[i]);
-
-                await _memberRepository.UpdateMemberProperties(memberId, membersProperties[i]);
-
-                if (membersSubscriptions.ContainsKey(i))
-                {
-                    MemberSubscription subscription = new MemberSubscription(memberId, membersSubscriptions[i].Type, membersSubscriptions[i].ExpiryDate);
-                    await _memberRepository.UpdateMemberSubscription(subscription);
-                }
-            }
-        }
-
-        public async Task<MemberImage> RotateMemberImage(Guid currentMemberId, Guid memberId, int degrees)
-        {
-            Member member = await GetMember(currentMemberId, memberId);
-
-            MemberImage rotated = await _memberService.RotateMemberImage(member.Id, degrees);
+            await _memberService.RotateMemberImage(member.Id, degrees);
 
             _cacheService.RemoveVersionedItem<MemberImage>(memberId);
-
-            return rotated;
         }
-
-        public async Task<MemberImage> UpdateMemberImage(Guid currentMemberId, Guid memberId, UpdateMemberImage image)
-        {
-            Member member = await GetMember(currentMemberId, memberId, true);
-
-            MemberImage updated = await _memberService.UpdateMemberImage(member.Id, image);
-
-            _cacheService.RemoveVersionedItem<MemberImage>(memberId);
-
-            return updated;
-        }
-
+        
         public async Task<ServiceResult> UpdateMemberSubscription(Guid currentMemberId, Guid memberId,
             UpdateMemberSubscription subscription)
         {
-            Member member = await GetMember(currentMemberId, memberId);
+            Member? member = await GetMember(currentMemberId, memberId);
             if (member == null)
             {
                 return ServiceResult.Failure("Member not found");
@@ -200,29 +135,10 @@ namespace ODK.Services.Members
 
             return ServiceResult.Successful();
         }
-
-        public async Task<MemberSubscription> UpdateMemberSubscriptionOld(Guid currentMemberId, Guid memberId,
-            UpdateMemberSubscription subscription)
+        
+        private async Task<Member?> GetMember(Guid currentMemberId, Guid id, bool superAdmin)
         {
-            Member member = await GetMember(currentMemberId, memberId);
-
-            DateTime? expiryDate = subscription.Type == SubscriptionType.Alum ? new DateTime?() : subscription.ExpiryDate;
-
-            MemberSubscription update = new MemberSubscription(member.Id, subscription.Type, expiryDate);
-
-            AssertValidMemberSubscription(update);
-
-            await _memberRepository.UpdateMemberSubscription(update);
-
-            _cacheService.RemoveVersionedItem<MemberSubscription>(memberId);
-            _cacheService.RemoveVersionedCollection<Member>(member.ChapterId);
-
-            return update;
-        }
-
-        private async Task<Member> GetMember(Guid currentMemberId, Guid id, bool superAdmin)
-        {
-            Member member = await _memberRepository.GetMember(id, true);
+            Member? member = await _memberRepository.GetMember(id, true);
             if (member == null)
             {
                 throw new OdkNotFoundException();
@@ -239,25 +155,7 @@ namespace ODK.Services.Members
 
             return member;
         }
-
-        private void AssertValidMemberSubscription(MemberSubscription subscription)
-        {
-            if (!Enum.IsDefined(typeof(SubscriptionType), subscription.Type) || subscription.Type == SubscriptionType.None)
-            {
-                throw new OdkServiceException("Invalid type");
-            }
-
-            if (subscription.Type == SubscriptionType.Alum && subscription.ExpiryDate != null)
-            {
-                throw new OdkServiceException("Alum should not have expiry date");
-            }
-            
-            if (subscription.Type != SubscriptionType.Alum && subscription.ExpiryDate < DateTime.UtcNow.Date)
-            {
-                throw new OdkServiceException("Expiry date should not be in the past");
-            }
-        }
-
+        
         private ServiceResult ValidateMemberSubscription(MemberSubscription subscription)
         {
             if (!Enum.IsDefined(typeof(SubscriptionType), subscription.Type) || subscription.Type == SubscriptionType.None)
